[REFACTOR] memisahkan attribute tentang salary ke class FinancialInformation
